Fix failing tests

+++ c/src/core/util/bytes.h
@@ -64,7 +64,7 @@ typedef uint_fast8_t wlen_t;        /**< number of bytes within a word (min 1byt
 /** a byte array */
 typedef struct bytes {
   uint8_t* data; /**< the byte-data  */
-  size_t   len;  /**< the length of the array ion bytes */
+  uint32_t   len;  /**< the length of the array ion bytes */
 } bytes_t;
 
 /** a byte-buffer to attach byte-functions. */
@@ -73,7 +73,7 @@ typedef struct {
   bytes_t b;     /**< the bytes struct */
 } bytes_builder_t;
 
-bytes_t* b_new(const uint8_t* data, size_t len);              /**< allocates a new byte array with 0 filled */
+bytes_t* b_new(const uint8_t* data, uint32_t len);              /**< allocates a new byte array with 0 filled */
 void     b_print(const bytes_t* a);                           /**< prints a the bytes as hex to stdout */
 void     ba_print(const uint8_t* a, size_t l);                /**< prints a the bytes as hex to stdout */
 int      b_cmp(const bytes_t* a, const bytes_t* b);           /**< compares 2 byte arrays and returns 1 for equal and 0 for not equal*/
